local/requirements.txt
sounddevice
numpy
requests
opencv-python
pyzbar
watchdog
Pillow
langchain
langchain-core
openai
injective_py